matrix/requirements.txt
matrix-nio[e2e]
python-dotenv
